+++ relay-server/src/services/project_cache.rs
@@ -1186,7 +1186,7 @@ mod tests {
     use tokio::select;
     use uuid::Uuid;
 
-    use crate::services::processor::{ProcessingGroup, Ungrouped};
+    use crate::services::processor::ProcessingGroup;
     use crate::testutils::{empty_envelope, empty_envelope_with_dsn};
 
     use super::*;
@@ -1287,7 +1287,7 @@ mod tests {
                     empty_envelope(),
                     services.outcome_aggregator.clone(),
                     services.test_store.clone(),
-                    ProcessingGroup::Ungrouped(Ungrouped),
+                    ProcessingGroup::Ungrouped,
                 )
                 .unwrap();
             let message = ValidateEnvelope { envelope };
@@ -1379,7 +1379,7 @@ mod tests {
                     empty_envelope_with_dsn(dsn),
                     services.outcome_aggregator.clone(),
                     services.test_store.clone(),
-                    ProcessingGroup::Ungrouped(Ungrouped),
+                    ProcessingGroup::Ungrouped,
                 )
                 .unwrap();
 
@@ -1462,7 +1462,7 @@ mod tests {
                 empty_envelope_with_dsn(dsn),
                 services.outcome_aggregator.clone(),
                 services.test_store.clone(),
-                ProcessingGroup::Ungrouped(Ungrouped),
+                ProcessingGroup::Ungrouped,
             )
             .unwrap();