+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Set the span op on segments. ([#3082](https://github.com/getsentry/relay/pull/3082))
 - Skip profiles without required measurements. ([#3112](https://github.com/getsentry/relay/pull/3112))
 - Push metrics summaries to their own topic. ([#3045](https://github.com/getsentry/relay/pull/3045))
+- Add `user.sentry_user` computed field for the on demand metrics extraction pipeline. ([#3122](https://github.com/getsentry/relay/pull/3122))
 
 ## 24.1.2